data-engineering/zoomcamp/module-4/pyproject.toml
[project]
name = "module-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastparquet>=2024.11.0",
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-storage>=3.9.0",
    "polars>=1.38.1",
    "pyarrow>=23.0.0",
]
